pyproject.toml
[project]
name = "daily-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "garminconnect>=0.2.25",
    "icecream>=2.1.4",
    "pandas>=2.2.3",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
]
